extract radio components into separate modules
